tibbisekreter-chat/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai==0.28.1
python-dotenv==1.0.0 